cleanup suggestion builder
